--- riot/templates/riot/champ.html
@@ -1,177 +1,0 @@
-{% load static %}
-
-<link rel="stylesheet" type="text/css" href="{% static 'riot/record.css' %}">
-
-{% block content %}
-<html lang="en">
-  <head>
-      <title>{{ champ.summoner_name }}: {{ champ.champ_name }}</title>
-
-      <!-- Required meta tags -->
-      <meta charset="utf-8">
-      <meta name="viewport" content="width=device-width, initial-scale=1">
-
-      <!-- Bootstrap CSS -->
-      <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
-      <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.5.0/font/bootstrap-icons.css">
-      
-  </head>
-
-
-  <body>
-    {% include "riot/include/navbar.html" %}
-    <main>
-      <h1>RIOT API</h1>
-
-      <p><strong>{{ champ.summoner_name }}</strong> </p>    
-      <p>Summoner ID: <strong>{{ champ.summonerId }}</strong></p>    
-      <p>Account ID: <strong>{{ champ.accountId }}</strong></p>    
-      <p>Champion name: <strong>{{ champ.champ_name }}</strong></p>
-      <p>Champion id: <strong>{{ champ.championId }}</strong></p>
-      <p>Champion mastery: <strong>{{ champ.championLevel }}</strong></p>
-      <p>Champion points: <strong>{{ champ.championPoints }}</strong></p>
-      <p>Last time: <strong>{{ champ.last_time }}</strong></p>
-
-    </main>
-
-      <div class="container" style="border:2px solid black;">
-        <br>
-        {% for game in game_info_list %}
-          <span class="anchor" id="{{forloop.counter}}"></span>
-          <div class="container" style="border:1px solid rgb(64, 97, 71);">
-            <br>
-            <h4><a href="/{{champ.server}}/{{champ.summonerName}}/{{champ.champ_name}}/info">{{ champ.champ_name }}</a> {{game.position}} {{game.date}} <a href="/{{champ.server}}/{{game.game_id}}/info">{{ game.game_id}}</a></h4>
-
-            <form action="#{{forloop.counter}}" method="post">
-            {% csrf_token %}
-            <button name="load" type="submit" value="{{game.game_id}}" class="btn btn-info">Show game summary</button>
-            </form>
-            <br>
-              
-            <div id="load_more">
-              {% if game_data.success %}
-                {% if game_data.game_id == game.game_id %}
-                  <table class="table table-bordered">
-                    <thead>
-                        <tr>
-                            <th scope="col">Game ID</th>
-                            <th scope="col">Server</th>
-                            <th scope="col">When</th>
-                            <th scope="col">Duration</th>
-                            <th scope="col">Game Mode</th>
-                            <th scope="col">Patch</th>
-                        </tr>
-                    </thead>
-          
-                    <tbody>
-                        <tr>
-                            <td>{{ game_data.gameId }}</td>
-                            <td>{{ game_data.platformId }}</td>
-                            <td>{{ game_data.gameCreation }}</td>
-                            <td>{{ game_data.gameDuration }}</td>
-                            <td>{{ game_data.gameMode }}</td>
-                            <td>{{ game_data.gameVersion }}</td>
-                        </tr>
-                    </tbody>
-          
-                    <thead>
-                        <tr>
-                            <th scope="col">{{ game_data.teams.0.win }} (Blue Team)</th>
-                            <th scope="col">Rank</th>
-                            <th scope="col">KDA</th>
-                            <th scope="col">Damage</th>
-                            <th scope="col">CS</th>
-                            <th scope="col">Vision Score</th>
-                        
-                        </tr>
-                    </thead>
-                  
-                    <tbody>
-                        {% for player in game_data.participants|slice:":5" %}
-                        <tr >
-                            <td>{{ player.champion_name }}/<a href="/{{champ.server}}/{{player.summoner_name}}"> {{ player.summoner_name }}</a></td>
-                            <td>{{ player.tier }}</a></td>
-                            <td>{{ player.stats.kills }}/{{ player.stats.deaths }}/{{ player.stats.assists }}</td>
-                            <td>{{ player.stats.totalDamageDealtToChampions }}</td>
-                            <td>{{ player.stats.totalMinionsKilled }}</td>
-                            <td>{{ player.stats.visionScore }}</td>
-                        </tr>
-                        {% endfor %}
-                    </tbody>
-          
-                    <thead>
-                        <tr>
-                            <th scope="col">{{ game_data.teams.1.win }} (Red Team)</th>
-                            <th scope="col">Rank</th>
-                            <th scope="col">KDA</th>
-                            <th scope="col">Damage</th>
-                            <th scope="col">CS</th>
-                            <th scope="col">Vision Score</th>
-                        
-                        </tr>
-                    </thead>
-                  
-                    <tbody>
-                        {% for player in game_data.participants|slice:"5:" %}
-                          <tr>
-                            <td>{{ player.champion_name }}/<a href="/{{champ.server}}/{{player.summoner_name}}"> {{ player.summoner_name }}</a></td>
-                            <td>{{ player.tier }}</a></td>
-                            <td>{{ player.stats.kills }}/{{ player.stats.deaths }}/{{ player.stats.assists }}</td>
-                            <td>{{ player.stats.totalDamageDealtToChampions }}</td>
-                            <td>{{ player.stats.totalMinionsKilled }}</td>
-                            <td>{{ player.stats.visionScore }}</td>
-                          </tr>
-                        {% endfor %}
-                    </tbody>
-                  </table>
-                  <br>
-                {% endif %}
-              {% endif %}
-            </div>
-          </div>
-          <br>
-        {% endfor %}
-      </div>
-      <br>
-
-      {% if games.has_other_pages %}
-      <nav aria-label="Page navigation example">
-        <ul class="pagination justify-content-center">
-          {% if games.has_previous %}
-            <li class="page-item">
-              <a class="page-link" href="?page={{ games.previous_page_number }}">&laquo;</a>
-            </li>
-          {% else %}
-            <li class="page-item disabled">
-              <a class="page-link">&laquo;</a>
-            </li>
-          {% endif %}
-
-          {% for i in games.paginator.page_range %}
-            {% if games.number == i %}
-              <li class="page-item active">
-                <a class="page-link"> {{ i }}</a>
-              </li>
-            {% else %}
-              <li class="page-item"><a class="page-link" href="?page={{ i }}">{{ i }}</a></li>
-            {% endif %}
-          {% endfor %}
-
-          {% if games.has_next %}
-            <li class="page-item">
-              <a class="page-link" href="?page={{ games.next_page_number }}">&raquo;</a>
-            </li>
-          {% else %}
-            <li class="page-item disabled">
-              <a class="page-link">&raquo;</a>
-            </li>
-          {% endif %}
-        </ul>
-      </nav>
-      {% endif %}
-
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
-
-  </body>
-</html>
-{% endblock %}
