+++ riot/templates/riot/include/user-profile-page.html
@@ -1,10 +1,10 @@
 {% load el_pagination_tags %}
-{% lazy_paginate game_info_list %}
-{% for game in game_info_list %}
+{% lazy_paginate game_summary_list %}
+{% for game in game_summary_list %}
     <span class="anchor" id="{{forloop.counter}}"></span>
     <div class="container" style="border:1px solid rgb(64, 97, 71);">
     <br>
-    <h4><a href="/{{user_account_info.server}}/{{ranked_stats.summonerName}}/{{game.champion_name}}/info">{{ game.champion_name }}</a> {{game.position}} {{game.date}} {{ game.game_id}}</h4>
+    <h4><a href="/{{user_account_info.server}}/{{user_account_info.name}}/{{game.champion_name}}">{{ game.champion_name }}</a> {{game.position}} {{game.date}} {{ game.game_id}}</h4>
 
     <form action="#{{forloop.counter}}" method="post">
     {% csrf_token %}
